feat: support onlcr mode
This adds support for recognizing the ONLCR mode when the terminal is
not a TTY. We need that in Wish v2 so that emulated ptys work as
expected and mouse movements don't break when an emulated pty is being
used. Since emulated ptys are not TTYs, we cannot set them to raw
mode which means we can't disable the default ONLCR mode.

Signed-off-by: Ayman Bagabas <ayman.bagabas@gmail.com>