fix: input: introduce esc timeout and notify channels
Terminal input sequences are complicated. When we receive an ESC byte,
it doesn't mean we have a complete sequence yet. It might be part of an
incomplete sequence. Thus, we need to introduce a timeout interval
mechanism to detect if we have a complete sequence or not.

For example, we might get "\x1b[A" (arrow up key) as 3 separate reads
one byte at a time. Introducing a timeout allows us to detect that and
be able to distinguish between a complete sequence, a single ESC byte
(escape key), or an incomplete sequence.