fix: renderer: make sure we correctly handle inline scrollHeight
When we're in inline mode, we need to ensure we use '\n' when the height
changes to make sure we scroll the terminal screen correctly. Otherwise,
the output might not scroll the screen. This is necessary because when
we run in inline mode and with relative cursor movement, we might not
have the full height of the application available to us in the screen
like when the current cursor position is at the bottom and we need to
scroll up to make space for new content.